1.去掉SceneType枚举，改成int，解除Core程序集对SceneType的依赖 2.SceneType不再使用Flag形式，多个Scene订阅可以加多个标签

+++ Unity/Assets/Scripts/Hotfix/Client/Module/Message/NetComponentOnReadInvoker_NetClient.cs
@@ -2,7 +2,7 @@
 
 namespace ET.Client
 {
-    [Invoke((long)SceneType.NetClient)]
+    [Invoke(SceneType.NetClient)]
     public class NetComponentOnReadInvoker_NetClient: AInvokeHandler<NetComponentOnRead>
     {
         public override void Handle(NetComponentOnRead args)
