+++ Unity/Assets/Scripts/Hotfix/Server/Demo/LockStep/Match/FiberInit_Match.cs
@@ -2,7 +2,7 @@
 
 namespace ET.Server
 {
-    [Invoke((long)SceneType.Match)]
+    [Invoke(SceneType.Match)]
     public class FiberInit_Match: AInvokeHandler<FiberInit, ETTask>
     {
         public override async ETTask Handle(FiberInit fiberInit)
