+++ Unity/Assets/Scripts/Hotfix/Server/Demo/StateSync/Gate/FiberInit_Gate.cs
@@ -2,7 +2,7 @@
 
 namespace ET.Server
 {
-    [Invoke((long)SceneType.Gate)]
+    [Invoke(SceneType.Gate)]
     public class FiberInit_Gate: AInvokeHandler<FiberInit, ETTask>
     {
         public override async ETTask Handle(FiberInit fiberInit)
