+++ Unity/Assets/Scripts/Hotfix/Server/Module/ActorLocation/FiberInit_Location.cs
@@ -2,7 +2,7 @@
 
 namespace ET.Server
 {
-    [Invoke((long)SceneType.Location)]
+    [Invoke(SceneType.Location)]
     public class FiberInit_Location: AInvokeHandler<FiberInit, ETTask>
     {
         public override async ETTask Handle(FiberInit fiberInit)
