+++ Unity/Assets/Scripts/Hotfix/Server/Module/Router/FiberInit_Router.cs
@@ -2,7 +2,7 @@
 
 namespace ET.Server
 {
-    [Invoke((long)SceneType.Router)]
+    [Invoke(SceneType.Router)]
     public class FiberInit_Router: AInvokeHandler<FiberInit, ETTask>
     {
         public override async ETTask Handle(FiberInit fiberInit)
